requirements (1).txt
fastapi==0.68.1
uvicorn==0.15.0
redis==4.0.2
pyjwt==2.3.0